orphaned_vms.sh: initial commit
orphaned_vms.sh is a script that searches through the current VMs (as
reported by nova) and retrieves their own.  Then, it cross checks that
list against a tenant listing  (as reported by keystone).  Any VMs whose
owner does not exist in keystone is assumed to be orphaned. This can
happen if a tenant is deleted while it still has VMs online.

Change-Id: I880b66e6d303e3348ac1d7fde1762633ae9ac07a