orphaned_volumes.sh: initial commit
orphaned_volumes.sh is a script that gets a list of all volumes and
their owner (as reported by cinder) and compares it to a list of all
tenants (as reported by keystone).  If any volume has an owner who does
not exist in keystone, it is returned in the output.  Orphaned cinder
volumes can occur when a tenant is deleted but still had volumes
provisioned.

Change-Id: I3e20826be54595b4a9a35aaac95be881ce658fa0