Update for Juno. Usage is per user = big changes
The upstream version was a good example, but nova has got a lot more
complex since then. These changes are a bit hacky, but try to fix quota
usage per user for a given project.

Also the code to check if the project quota usage matches reality only
checks the per project totals, it might be that the total is right, but
the values in the quota_usages table are wrong, but cancel each other
out - maybe we can just live with that!