Update aut-fix-quota tool for Juno
I've ported this tool to work with Juno and thought I'd share this for
all to use.

We've done basic functional testing, an internal code review and used it
against our production cloud. I would encourage further review and
testing, but this should be usable for others.

The backend changes to Juno required more changes that I would have
liked and the code does throw the odd exception (handled by the code)
when run against every project in our production system, but this hasn't
caused any issues for us.

Change-Id: I3913251ede4949149e63d1337bd92dd836f98763